pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rosa-os"
version = "1.0.0"
description = "ROSA OS — Hybrid AI Assistant Platform"
requires-python = ">=3.11"
dependencies = [
    # Existing
    "openai>=2.26.0",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    # New: ROSA Core backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    # New: Memory layer
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    # New: Tools
    "lxml>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
